[MOVE] Token manager from util to data package
