+++ app/src/main/java/dev/hossain/trmnl/ui/display/TrmnlMirrorDisplayScreen.kt
@@ -56,12 +56,12 @@ import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import dev.hossain.trmnl.R
 import dev.hossain.trmnl.data.ImageMetadataStore
+import dev.hossain.trmnl.data.TokenManager
 import dev.hossain.trmnl.di.AppScope
 import dev.hossain.trmnl.ui.FullScreenMode
 import dev.hossain.trmnl.ui.refreshlog.DisplayRefreshLogScreen
 import dev.hossain.trmnl.ui.settings.AppSettingsScreen
 import dev.hossain.trmnl.util.CoilRequestUtils
-import dev.hossain.trmnl.util.TokenManager
 import dev.hossain.trmnl.util.nextRunTime
 import dev.hossain.trmnl.work.TrmnlImageUpdateManager
 import dev.hossain.trmnl.work.TrmnlWorkScheduler
